+++ src/main/java/no/nsd/qddt/domain/category/Category.java
@@ -423,7 +423,7 @@ public class Category extends AbstractEntityAudit  implements Comparable<Categor
     }
 
     @Override
-    public void makeNewCopy(Long revision){
+    public void makeNewCopy(Integer revision){
         // Copying a simple Category doesn't make any sense... skipping...
         hasRun = (getHierarchyLevel() == HierarchyLevel.ENTITY);
         if (hasRun) return;
