+++ src/main/java/no/nsd/qddt/domain/publication/audit/PublicationAuditService.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 /**
  * @author Dag Østgulen Heradstveit
  */
-public interface PublicationAuditService extends BaseServiceAudit<Publication, UUID, Long> {
+public interface PublicationAuditService extends BaseServiceAudit<Publication, UUID, Integer> {
 
-    Page<Revision<Long, Publication>> findRevisionByIdAndChangeKindNotIn(UUID id, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
+    Page<Revision<Integer, Publication>> findRevisionByIdAndChangeKindNotIn(UUID id, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
 }
