+++ src/main/java/no/nsd/qddt/domain/study/audit/StudyAuditRepository.java
@@ -15,7 +15,7 @@ import java.util.UUID;
  * @author Dag Østgulen Heradstveit
  */
 @Repository
-interface StudyAuditRepository extends RevisionRepository<Study, UUID, Long> {
+interface StudyAuditRepository extends RevisionRepository<Study, UUID, Integer> {
 
-    Page<Revision<Long,Study>> findRevisionsByIdAndChangeKindNotIn(UUID uuid, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
+    Page<Revision<Integer,Study>> findRevisionsByIdAndChangeKindNotIn(UUID uuid, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
 }
