+++ src/main/java/no/nsd/qddt/domain/surveyprogram/SurveyProgram.java
@@ -128,7 +128,7 @@ public class SurveyProgram extends AbstractEntityAudit implements Authorable,Arc
     }
 
     @Override
-    public void makeNewCopy(Long revision){
+    public void makeNewCopy(Integer revision){
         if (hasRun) return;
         super.makeNewCopy(revision);
         getStudies().forEach(s->s.makeNewCopy(revision));
